% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grList_to_df}
\alias{grList_to_df}
\title{Convert grange list to dataframe}
\usage{
grList_to_df(grange.list, anno.id = "cell.population")
}
\arguments{
\item{anno.id}{}
}
\description{
Convert grange list to dataframe
}
\keyword{internal}
