% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBigwig_summary_SS.R
\name{multiBigwig_summary_SS}
\alias{multiBigwig_summary_SS}
\title{Summarize signal from BigWig files over genomic regions}
\usage{
multiBigwig_summary_SS(
  data_table,
  data_frame_of_regions,
  summary_type = "mean",
  parallel = TRUE
)
}
\arguments{
\item{data_table}{A data.frame with columns \code{bw_path} and \code{sample_id}}

\item{data_frame_of_regions}{A data.frame with columns \code{chr}, \code{start}, \code{end}}

\item{summary_type}{Summary type: "mean", "median", "min", or "max" (default = "mean")}

\item{parallel}{Logical. Use parallel processing? Default is TRUE.}
}
\value{
A data.frame with signal summaries per region per sample
}
\description{
Computes summary statistics (mean, median, min, or max)
across genomic regions from multiple BigWig files.
Inspired by deeptools' multiBigwigSummary, bedtoolsr (Phanstiel lab)
}
