% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_MACS2_rank.R
\name{peak_annotation_rank}
\alias{peak_annotation_rank}
\title{A Peak_annotation_rank function}
\usage{
peak_annotation_rank(
  peak_table,
  group_by_column_name,
  rank_columns = c("peak.magnitude", "fold.change")
)
}
\arguments{
\item{group_by_column_name}{}
}
\description{
This function allows you to extract the peaks for a group by category along with its MACS2 rank from the CERP tsv file and places the output in a new directory called peak_subset_files with two files per group,
one as is and one sorted using bedtools in two directories
}
\keyword{MACS2}
