% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_coverage_rank.R
\name{peak_coverage}
\alias{peak_coverage}
\title{A peak_coverage_rank function}
\usage{
peak_coverage(peak_table.list, peak_table, bigwig.dir)
}
\arguments{
\item{bigwig.dir}{}
}
\description{
The Peak_coverage() function calculates coverage for a groups' peak  across background groups list provided using multiBigWigSummary..
and Peak_coverage_rank() function ranks (ascending) the coverage proportion per peak of a group across groups
Needs the bigwig files
}
\keyword{multiBigWigSummary}
