% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_intersect_rank.R
\name{peak_intersect}
\alias{peak_intersect}
\title{A Peak_intersect_rank function}
\usage{
peak_intersect(peak_table.list)
}
\arguments{
\item{peak_table.list}{}
}
\description{
This Peak_intersect() function intersects each group peaks by background_group peak files using bedtools and
Peak_intersect_rank() ranks for specificity (group peaks with low intersection). The results are stored in peak_intersect_rank folder.
}
\keyword{bedtools}
\keyword{intersect}
