% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_intersect_rank.R
\name{Peak_intersect_rank}
\alias{Peak_intersect_rank}
\title{A Peak_intersect_rank function}
\usage{
Peak_intersect_rank(tsv_file_df, group_by_column_name, group, background_group)
}
\arguments{
\item{tsv_file_df}{CERP tsv_file_df REQUIRED}

\item{group}{unique group in the tsv_file for that needs to be ranked}

\item{background_group}{list of groups from tsv_file with which the group peaks should be intersected}

\item{group_by_column}{column name in the tsv_file that has the groups}
}
\description{
Peak_intersect_rank() ranks for specificity (peaks of a group having low intersection with other groups).
}
