% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_MACS2_rank.R
\name{Peak_MACS2_rank}
\alias{Peak_MACS2_rank}
\title{A Peak_MACS2_rank function}
\usage{
Peak_MACS2_rank(tsv_file_df, group_by_column_name, group)
}
\arguments{
\item{tsv_file_df}{CERP tsv_file_df REQUIRED}

\item{group}{unique group in the tsv_file for which the peak rank file needs to be created, for eg. Pvalb}

\item{group_by_column}{column name in the tsv_file that has the groups}
}
\description{
This function allows you to extract the peaks for a group by category along with its MACS2 rank from the CERP tsv file and places the output in a new directory called peak_subset_files with two files per group,
one as is and one sorted using bedtools in two directories
}
\examples{
Peak_MACS2_rank()

}
\keyword{MACS2}
