% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_magnitude_rank.R
\name{Peak_magnitude_rank}
\alias{Peak_magnitude_rank}
\title{Peak signal magnitude scaling and ranking}
\usage{
Peak_magnitude_rank(multiBigwig_summary_df, bw_table)
}
\arguments{
\item{multiBigwig_summary_df}{A data.frame from multiBigwig_summary_SS()}

\item{bw_table}{A data.frame with columns \code{bw_path} and \code{sample_id}}
}
\value{
A data.frame of the same regions with magnitude ranks per sample
}
\description{
For a set of genomic regions, this function:
\itemize{
\item Takes already summarized signal magnitudes from BigWig files
\item Scales those signals per sample to \link{0–1}
\item Ranks the scaled magnitudes per region per sample (descending)
}
}
