% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRankeR.R
\name{Peak_RankeR}
\alias{Peak_RankeR}
\title{Peak_RankeR function}
\usage{
Peak_RankeR(
  tsv_file_df,
  group_by_column_name,
  background_group,
  bw_table,
  rank_sum,
  weights
)
}
\arguments{
\item{tsv_file_df}{CERP tsv_file_df REQUIRED}

\item{background_group}{list of groups from tsv_file with which the group peaks should be intersected}

\item{bw_table}{table of sample name and bigwig files}

\item{rank_sum}{if the rank sum (0 - 1) should be provided in the output}

\item{weights}{vector of the features MACS2, intersect, coverage in that order}

\item{group_by_column}{column name in the tsv_file that has the groups}
}
\description{
This function calculates the final rank for group of interest against background group
}
