% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRankeR.R
\name{PeakRankeR}
\alias{PeakRankeR}
\title{A Peak_RankeR function}
\usage{
PeakRankeR(peak_table, group_by_column_name, bigwig.dir)
}
\arguments{
\item{rank_df}{data_frame which has the peaks for a specific group and newly added rank columns generated using PeakRankeR package}
}
\description{
This function calculates the final rank for each peak in tsv file given the columns of *ranks produced using PeakRankeR package
}
\keyword{PeakRankeR}
