% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRankeR.R
\name{range01}
\alias{range01}
\title{A Peak_RankeR function}
\usage{
range01(x)
}
\arguments{
\item{rank_df}{data_frame which has the peaks for a specific group and newly added rank columns generated using PeakRankeR package}
}
\description{
This function calculates the final rank for each peak in tsv file given the columns of *ranks produced using PeakRankeR package
}
\examples{
Peak_RankeR()
range01()

}
\keyword{PeakRankeR}
