% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBigwig_summary_SS.R
\name{multiBigwig_summary_SS}
\alias{multiBigwig_summary_SS}
\title{Enrichments at genomics regions (Adapted from ALPS)}
\usage{
multiBigwig_summary_SS(
  data_table,
  summary_type = "mean",
  data_frame_of_regions,
  parallel = TRUE
)
}
\arguments{
\item{data_table}{a dataframe that contains \code{bw_path}, \code{sample_id}.
\code{sample_id} ids will be used in the final result matrix}

\item{summary_type}{whether to calculate mean, median, min or max
for each genomic region in within consensus peak-set
from the bigwig files. Default is \code{mean}
Added by SS}

\item{data_frame_of_regions}{df containing chr, start, end for group of interest}

\item{parallel}{logical. Whether to parallelize the calculation process,
default is \code{TRUE}}
}
\value{
data.frame of enrichments within given genomic regions
}
\description{
\code{multiBigwig_summary} is a function to calculate
enrichments from a set of given bigwig files
and a set of genomics regions.
This function is similar to
\code{deeptools multiBigwigSummary} python package.
}
