% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_coverage_rank.R
\name{Peak_coverage}
\alias{Peak_coverage}
\title{A Peak_coverage_rank function}
\usage{
Peak_coverage(group, background_group)
}
\arguments{
\item{group}{unique group  in the tsv_file for which the coverage across groups needs to be extracted}

\item{background_group}{list of groups from tsv_file from which coverage needs to be extracted for the gorup peaks}

\item{group_coverage_df}{result of multiBigWigSummary for group of interest}

\item{group_bigwig_file_name}{group of interest bigwig file name}

\item{group_macs2_intersect_rank_df}{df of ranks calculated for other features}
}
\description{
The Peak_coverage() function calculates coverage for a groups' peak  across background groups list provided using multiBigWigSummary..
and Peak_coverage_rank() function ranks (ascending) the coverage proportion per peak of a group across groups
Needs the bigwig files
}
\examples{
Peak_coverage()
Peak_coverage_rank()
}
\keyword{multiBigWigSummary}
