% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_coverage_rank.R
\name{Peak_coverage_rank}
\alias{Peak_coverage_rank}
\title{A Peak_coverage_rank function}
\usage{
Peak_coverage_rank(
  tsv_file_df,
  group_by_column,
  group,
  background_group,
  bw_table
)
}
\arguments{
\item{tsv_file_df}{CERP tsv_file_df REQUIRED}

\item{group_by_column}{column name in the tsv_file that has the groups}

\item{group}{unique group in the tsv_file that needs to be ranked}

\item{background_group}{list of groups from tsv_file with which the group peaks should be intersected}

\item{bw_table}{table of sample name and bigwig files}
}
\description{
Peak_coverage_rank() function ranks (descending) the coverage proportion per peak of a group across groups
Needs the bigwig files
}
