% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_intersect_rank.R
\name{Peak_intersect}
\alias{Peak_intersect}
\title{A Peak_intersect_rank function}
\usage{
Peak_intersect(group, background_group)
}
\arguments{
\item{group}{unique group in the tsv_file for which the ranked peak intersect file needs to be created}

\item{background_group}{list of groups from tsv_file with which the group peaks should be intersected}

\item{group_intersected_file_df}{the intersected file for group of interest produced by bedtools}
}
\description{
This Peak_intersect() function intersects each group peaks by background_group peak files using bedtools and
Peak_intersect_rank() ranks for specificity (group peaks with low intersection). The results are stored in peak_intersect_rank folder.
}
\examples{
Peak_intersect()
Peak_intersect_rank()
}
\keyword{bedtools}
\keyword{intersect}
