% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_MACS2_rank.R
\name{Peak_MACS2_rank}
\alias{Peak_MACS2_rank}
\title{A Peak_MACS2_rank function}
\usage{
Peak_MACS2_rank(tsv_file_df, group_by_column_name)
}
\arguments{
\item{tsv_file_df}{CERP tsv_file_df REQUIRED}

\item{group_by_column}{column name in the tsv_file that has the groups}
}
\description{
Add rank to the tsv file based on peak magnitude given by MACS2
}
\examples{
Peak_MACS2_rank()

}
\keyword{MACS2}
